Implement tanersener/mobile-ffmpeg integration for devices that can support it
