+++ mobile-ffmpeg-full-4.4/build.gradle
@@ -1,0 +1,2 @@
+configurations.maybeCreate("default")
+artifacts.add("default", file('mobile-ffmpeg-full-4.4.aar'))
