+++ settings.gradle
@@ -1,1 +1,2 @@
+include ':mobile-ffmpeg-full-4.4'
 include ':app', ':sdk', ':settingsutil', ':licensehelper', ':streaming'
