+++ streaming/src/main/AndroidManifest.xml
@@ -1,6 +1,8 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="org.btelman.controlsdk.streaming">
+    xmlns:tools="http://schemas.android.com/tools"
+    package="org.btelman.controlsdk.streaming">
     <uses-permission android:name="android.permission.CAMERA"/>
     <uses-feature android:name="android.hardware.camera"/>
     <uses-feature android:name="android.hardware.camera.autofocus"/>
+    <uses-sdk tools:overrideLibrary="com.arthenica.mobileffmpeg"/>
 </manifest>
