+++ streaming/src/main/java/org/btelman/controlsdk/streaming/video/retrievers/DummyRetriever.kt
@@ -1,0 +1,14 @@
+package org.btelman.controlsdk.streaming.video.retrievers
+
+import org.btelman.controlsdk.streaming.models.ImageDataPacket
+
+/**
+ * Return the same packet. Nothing actually gets done here
+ */
+class DummyRetriever : BaseVideoRetriever() {
+    private val imageDataPacket = ImageDataPacket(null)
+    override fun grabImageData(): ImageDataPacket? {
+        imageDataPacket.timecode = System.currentTimeMillis()
+        return imageDataPacket
+    }
+}
